Feat/cookie auth (#107)
* cookie-based auth

* big changes incoming

* FIx auth tests

* fix auth

* add helpers

* delete user improvements

* fix all tests

* ChangePassword.cs move to users tag

* fix frontend

+++ server/src/Application/ServiceRecords/Delete/DeleteServiceRecordCommandHandler.cs
@@ -2,7 +2,6 @@
 using Application.Abstractions.Data;
 using Application.Abstractions.Messaging;
 using Application.Core;
-using Application.Vehicles;
 using Microsoft.EntityFrameworkCore;
 
 namespace Application.ServiceRecords.Delete;
@@ -13,26 +12,16 @@ internal sealed class DeleteServiceRecordCommandHandler(IApplicationDbContext db
     {
         var serviceRecord = await dbContext.ServiceRecords
             .Include(sr => sr.Vehicle)
-            .FirstOrDefaultAsync(sr => sr.Id == request.ServiceRecordId && sr.VehicleId == request.VehicleId, cancellationToken);
+            .FirstOrDefaultAsync(sr =>
+                    sr.Id == request.ServiceRecordId &&
+                    sr.VehicleId == request.VehicleId,
+                cancellationToken);
 
-        if (serviceRecord is null)
-            return Result.Failure(ServiceRecordErrors.NotFound(request.ServiceRecordId));
+        if (serviceRecord?.Vehicle is null || serviceRecord.Vehicle.UserId != userContext.UserId)
+            return Result.Failure(ServiceRecordErrors.NotFound);
 
-        if (serviceRecord.Vehicle is null)
-            return Result.Failure(VehicleErrors.NotFound(request.VehicleId));
-
-        if (serviceRecord.Vehicle.UserId != userContext.UserId)
-            return Result.Failure(ServiceRecordErrors.Unauthorized);
-
-        try
-        {
-            dbContext.ServiceRecords.Remove(serviceRecord);
-            await dbContext.SaveChangesAsync(cancellationToken);
-        }
-        catch (DbUpdateException)
-        {
-            return Result.Failure(ServiceRecordErrors.DeleteFailed(request.ServiceRecordId));
-        }
+        dbContext.ServiceRecords.Remove(serviceRecord);
+        await dbContext.SaveChangesAsync(cancellationToken);
 
         return Result.Success();
     }
